Attempt to fix issue with ghost immediately changing room.
Before, we had a bug where an angry ghost would leave its ruined room
and then when it hit the next room, immediately pivot and go to a third
room. This is strange.

I tried preventing the "tell the ghost to move" function from running if
the ghost has a destination in mind already, and also I made it so we only
run the function if the dirtiness of the room has changed.

I haven't seen it again, and I think these are safe changes, so here they are.