Ghosts don't get angry while in the summoning room.
This is sort of an edge case handling, but now ghosts can stay in the
summoning room and they'll be alright.

We may want, at some point, to instead just randomly reassign them.

Also, changes the graphics on the ghosts so that they are unhappy
while still in the state where you can redirect them, but once they
enter unfixable state they change to angry face. I think this is
probably cleaner?